A somewhat functional and tested `check_perfdata`
For an employer, I once did a Bash version of this idea which had a
couple of downsides:

  1. It was agonizingly slow (because of (my inaptitude with) Bash).
  2. It didn't interpret Nagios' perfdata “spec” correctly.

This seemed like a simple-enough program to teach myself some Rust. At
this point, there's quite a of (semi-)dead code, though, partially
because the code base seems incapable of making up its mind about
whether it wants to be a library or just a binary and its supporting
code. Forgive me the libraritis (for now); I'm a mere scripter playing
at engineering.